Split each worker into its own module and struct
