+++ relayer/src/worker.rs
@@ -1,18 +1,8 @@
-use std::{fmt, thread, time::Duration};
+use std::fmt;
 
-use anomaly::BoxError;
-use crossbeam_channel::Receiver;
-use tracing::{debug, error, error_span, info, trace, warn};
+use tracing::{debug, error, info};
 
-use ibc::{events::IbcEvent, ics02_client::events::UpdateClient};
-
-use crate::{
-    chain::handle::ChainHandlePair,
-    foreign_client::{ForeignClient, ForeignClientError, MisbehaviourResults},
-    link::{Link, LinkParameters},
-    object::{Client, Object, UnidirectionalChannelPath},
-    util::retry::{retry_with_index, RetryResult},
-};
+use crate::{chain::handle::ChainHandlePair, object::Object};
 
 mod handle;
 pub use handle::WorkerHandle;
@@ -20,30 +10,21 @@ pub use handle::WorkerHandle;
 mod cmd;
 pub use cmd::WorkerCmd;
 
-mod retry_strategy {
-    use crate::util::retry::{clamp_total, ConstantGrowth};
-    use std::time::Duration;
+mod client;
+pub use client::ClientWorker;
 
-    const MAX_DELAY: Duration = Duration::from_millis(500);
-    const DELAY_INCR: Duration = Duration::from_millis(100);
-    const INITIAL_DELAY: Duration = Duration::from_millis(200);
-    const MAX_RETRY_DURATION: Duration = Duration::from_secs(2);
-
-    pub fn uni_chan_path() -> impl Iterator<Item = Duration> {
-        let strategy = ConstantGrowth::new(INITIAL_DELAY, DELAY_INCR);
-        clamp_total(strategy, MAX_DELAY, MAX_RETRY_DURATION)
-    }
-}
+mod uni_chan_path;
+pub use uni_chan_path::UniChanPathWorker;
 
 /// A worker processes batches of events associated with a given [`Object`].
-pub struct Worker {
-    chains: ChainHandlePair,
-    rx: Receiver<WorkerCmd>,
+pub enum Worker {
+    Client(ClientWorker),
+    UniChanPath(UniChanPathWorker),
 }
 
 impl fmt::Display for Worker {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "[{} <-> {}]", self.chains.a.id(), self.chains.b.id(),)
+        write!(f, "[{} <-> {}]", self.chains().a.id(), self.chains().b.id(),)
     }
 }
 
@@ -60,20 +41,22 @@ impl Worker {
             object,
         );
 
-        let worker = Self { chains, rx };
-        let thread_handle = std::thread::spawn(move || worker.run(object));
+        let worker = match object {
+            Object::Client(client) => Self::Client(ClientWorker::new(chains, rx, client)),
+            Object::UnidirectionalChannelPath(path) => {
+                Self::UniChanPath(UniChanPathWorker::new(chains, rx, path))
+            }
+        };
 
+        let thread_handle = std::thread::spawn(move || worker.run());
         WorkerHandle::new(tx, thread_handle)
     }
 
     /// Run the worker event loop.
-    fn run(self, object: Object) {
-        let span = error_span!("worker loop", worker = %object.short_name());
-        let _guard = span.enter();
-
-        let result = match object {
-            Object::UnidirectionalChannelPath(path) => self.run_uni_chan_path(path),
-            Object::Client(client) => self.run_client(client),
+    fn run(self) {
+        let result = match self {
+            Self::Client(w) => w.run(),
+            Self::UniChanPath(w) => w.run(),
         };
 
         if let Err(e) = result {
@@ -83,142 +66,10 @@ impl Worker {
         info!("worker exits");
     }
 
-    fn run_client_misbehaviour(
-        &self,
-        client: &ForeignClient,
-        update: Option<UpdateClient>,
-    ) -> bool {
-        match client.detect_misbehaviour_and_submit_evidence(update) {
-            MisbehaviourResults::ValidClient => false,
-            MisbehaviourResults::VerificationError => {
-                // can retry in next call
-                false
-            }
-            MisbehaviourResults::EvidenceSubmitted(_) => {
-                // if evidence was submitted successfully then exit
-                true
-            }
-            MisbehaviourResults::CannotExecute => {
-                // skip misbehaviour checking if chain does not have support for it (i.e. client
-                // update event does not include the header)
-                true
-            }
-        }
-    }
-
-    /// Run the event loop for events associated with a [`Client`].
-    fn run_client(self, client: Client) -> Result<(), BoxError> {
-        let mut client = ForeignClient::restore(
-            &client.dst_client_id,
-            self.chains.a.clone(),
-            self.chains.b.clone(),
-        );
-
-        info!(
-            "running client worker & initial misbehaviour detection for {}",
-            client
-        );
-
-        // initial check for evidence of misbehaviour for all updates
-        let skip_misbehaviour = self.run_client_misbehaviour(&client, None);
-
-        info!(
-            "running client worker (misbehaviour and refresh) for {}",
-            client
-        );
-
-        loop {
-            thread::sleep(Duration::from_millis(600));
-
-            // Run client refresh, exit only if expired or frozen
-            if let Err(e @ ForeignClientError::ExpiredOrFrozen(..)) = client.refresh() {
-                error!("failed to refresh client '{}': {}", client, e);
-                continue;
-            }
-
-            if skip_misbehaviour {
-                continue;
-            }
-
-            if let Ok(WorkerCmd::IbcEvents { batch }) = self.rx.try_recv() {
-                trace!("client '{}' worker receives batch {:?}", client, batch);
-
-                for event in batch.events {
-                    if let IbcEvent::UpdateClient(update) = event {
-                        debug!("client '{}' updated", client);
-
-                        // Run misbehaviour. If evidence submitted the loop will exit in next
-                        // iteration with frozen client
-                        self.run_client_misbehaviour(&client, Some(update));
-                    }
-                }
-            }
-        }
-    }
-
-    /// Run the event loop for events associated with a [`UnidirectionalChannelPath`].
-    fn run_uni_chan_path(self, path: UnidirectionalChannelPath) -> Result<(), BoxError> {
-        let mut link = Link::new_from_opts(
-            self.chains.a.clone(),
-            self.chains.b.clone(),
-            LinkParameters {
-                src_port_id: path.src_port_id,
-                src_channel_id: path.src_channel_id,
-            },
-        )?;
-
-        // TODO: Do periodical checks that the link is closed (upon every retry in the loop).
-        if link.is_closed()? {
-            warn!("channel is closed, exiting");
-            return Ok(());
-        }
-
-        fn step(cmd: Option<WorkerCmd>, link: &mut Link, index: u64) -> RetryResult<(), u64> {
-            if let Some(cmd) = cmd {
-                let result = match cmd {
-                    WorkerCmd::IbcEvents { batch } => {
-                        // Update scheduled batches.
-                        link.a_to_b.update_schedule(batch)
-                    }
-                    WorkerCmd::NewBlock {
-                        height,
-                        new_block: _,
-                    } => link.a_to_b.clear_packets(height),
-                };
-
-                if let Err(e) = result {
-                    error!("{}", e);
-                    return RetryResult::Retry(index);
-                }
-            }
-
-            let result = link
-                .a_to_b
-                .refresh_schedule()
-                .and_then(|_| link.a_to_b.execute_schedule());
-
-            if let Err(e) = result {
-                error!("{}", e);
-                return RetryResult::Retry(index);
-            }
-
-            RetryResult::Ok(())
-        }
-
-        loop {
-            thread::sleep(Duration::from_millis(200));
-
-            let result = retry_with_index(retry_strategy::uni_chan_path(), |index| {
-                step(self.rx.try_recv().ok(), &mut link, index)
-            });
-
-            if let Err(retries) = result {
-                return Err(format!(
-                    "UnidirectionalChannelPath worker failed after {} retries",
-                    retries
-                )
-                .into());
-            }
+    fn chains(&self) -> &ChainHandlePair {
+        match self {
+            Self::Client(w) => &w.chains(),
+            Self::UniChanPath(w) => w.chains(),
         }
     }
 }
